Move token initializing logic to login
